moves helper functions into their own file
